Level loading buffer "fixed"
Game no longer uses a buffer for loading level.
Instead, when a client connects, if a game is in progress, the server
will tell the client what level to load.
This avoids a client joining very late the pain of loading and unloading
a level for each game already played.