fixed and added some sounds
I screwed up and lost the working files for most of the sounds.
So there is completely new handgun, smg and shotgun sounds but theyre
definitely better.
Added a sound when you pick up an item and for when you hit an enemy.
Would you guys also be able to make a new build so I can see how the
sounds feel in the game please?